+++ CHANGES.rst
@@ -2,6 +2,7 @@ v0.2.0
 ======
 - Signature of chempy.equilibria.roots, changed.
 - Added two new modules: chempy.util.table, chempy.util.graph
+- chempy.einstein_smoluchowski added
 - Reaction, ReactionSystems now expects stoichs etc. to be given wrt to Substance names.
 - Added chempy.chemistry.ArrheniusRate
 - EqSystemLog, EqSystemLin -> EqSystem, (NumSysLog, NumSysLin)